winch(aarch64): implify constant handling, part 1/N
This commit is the first step toward simplifying constant handling,
particularly for the aarch64 backend.

The main highlights in this patch are:

* Introduction of `ConstantPool` implemenetation on top of Cranlift
  primitives. The implemettaion is identical to the existing for x64,
  however, it's abstracted so that it can be easily consumed from any
  existing backend.
* Usage of the constant pool from aarch64, which simplifies the
  loading of constants, particularly floating point constants.

The main motivation behind this change is to _eventually_ detach the
implicit usage of the scatch register from constant loading as much as
possible, reducing the possibility of subtle bugs (like the one
described in https://github.com/bytecodealliance/wasmtime/pull/10829).

Note that I have a work-in-progress branch from where all these
changes are cherry picked from, to make everything easier to review.

A side effect of this change, is the improvement to the code
generation involving floating point constants. Prior to this change,
multiple moves were involved, with this patch, at most 1 move is
required and at worst one load is required.